pytest: Separate Session tests
